Patch Notes 1.1
I added a method called "sumOddIndex" that calculates the sum of elements in odd-numbered indexes from randomArray, and a println statement to print the returned value called "sumOdd" by "sumOddIndex". Then, I changed the title of the menu from "Welcome!!" to "MAIN MENU" cuz it was stupid to greet the user again and again :) lastly I moved the method called "calcAverDifference" up to above the main method.